removing old team picker view controller
